# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dance Analysis API",
    "description": "Video dance analysis with pose estimation via microservices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/routines": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routines",
        "description": "List all routines created by the current user.",
        "operationId": "list_routines_api_v1_routines_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutineResponse"
                  },
                  "type": "array",
                  "title": "Response List Routines Api V1 Routines Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Routine",
        "description": "Create a routine.\n\n- **title**: Routine title (1-255 characters)\n- **dance_id**: ID of the dance",
        "operationId": "create_routine_api_v1_routines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/routines/{routine_id}": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "Get Routine",
        "description": "Get a specific routine. Requires ownership.",
        "operationId": "get_routine_api_v1_routines__routine_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "routines"
        ],
        "summary": "Update Routine",
        "description": "Update a routine. Requires ownership.",
        "operationId": "update_routine_api_v1_routines__routine_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Routine Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Delete Routine",
        "description": "Delete a routine. Requires ownership.",
        "operationId": "delete_routine_api_v1_routines__routine_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{session_id}/videos": {
      "post": {
        "tags": [
          "session-videos"
        ],
        "summary": "Register Upload",
        "description": "Register a new video upload.\n\nReturns a presigned PUT URL for the client to upload the file directly.\n\n- **filename**: Name of the video file\n- **content_type**: MIME type (default: video/mp4)\n- **file_size**: Optional file size in bytes",
        "operationId": "register_upload_api_v1_sessions__session_id__videos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRegisterUpload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "session-videos"
        ],
        "summary": "List Videos",
        "description": "List videos for a session.\n\nDefault: only uploaded videos (visible to all session members).\nPass ?status=pending_upload to see your own pending uploads.",
        "operationId": "list_videos_api_v1_sessions__session_id__videos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VideoStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by video status",
              "title": "Status"
            },
            "description": "Filter by video status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoResponse"
                  },
                  "title": "Response List Videos Api V1 Sessions  Session Id  Videos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/videos/{video_id}": {
      "get": {
        "tags": [
          "session-videos"
        ],
        "summary": "Get Video",
        "description": "Get video metadata. Deleted videos return 404.",
        "operationId": "get_video_api_v1_sessions__session_id__videos__video_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session-videos"
        ],
        "summary": "Delete Video",
        "description": "Soft delete a video.\n\nSets status=deleted and migrates associated notes to session-level.",
        "operationId": "delete_video_api_v1_sessions__session_id__videos__video_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/videos/{video_id}/finalize": {
      "post": {
        "tags": [
          "session-videos"
        ],
        "summary": "Finalize Upload",
        "description": "Mark a video upload as complete.\n\nOnly the uploader can finalize their pending upload (idempotent).",
        "operationId": "finalize_upload_api_v1_sessions__session_id__videos__video_id__finalize_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/videos/{video_id}/stream-url": {
      "get": {
        "tags": [
          "session-videos"
        ],
        "summary": "Get Stream Url",
        "description": "Get a presigned stream URL. Only works for uploaded videos.",
        "operationId": "get_stream_url_api_v1_sessions__session_id__videos__video_id__stream_url_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoStreamUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoutineCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "dance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dance Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "dance_id"
        ],
        "title": "RoutineCreate",
        "description": "Request body for creating a routine."
      },
...
      "RoutineResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "dance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dance Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "dance_id",
          "id",
          "created_by",
          "created_at"
        ],
        "title": "RoutineResponse",
        "description": "Response body for routine data."
      },
...
      "RoutineUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "dance_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dance Id"
          }
        },
        "type": "object",
        "title": "RoutineUpdate",
        "description": "Request body for updating a routine."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoRegisterResponse": {
        "properties": {
          "video": {
            "$ref": "#/components/schemas/VideoResponse"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "video",
          "upload_url",
          "expires_at"
        ],
        "title": "VideoRegisterResponse",
        "description": "Response after registering a video upload (includes presigned URL)."
      },
      "VideoRegisterUpload": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "maxLength": 100,
            "title": "Content Type",
            "default": "video/mp4"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "VideoRegisterUpload",
        "description": "Request body for registering a new video upload."
      },
      "VideoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "routine_session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Session Id"
          },
          "uploaded_by": {
            "type": "string",
            "format": "uuid",
            "title": "Uploaded By"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "status": {
            "$ref": "#/components/schemas/VideoStatus"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uploaded_by",
          "storage_key",
          "status",
          "created_at"
        ],
        "title": "VideoResponse",
        "description": "Response body for video metadata."
      },
      "VideoStatus": {
        "type": "string",
        "enum": [
          "pending_upload",
          "uploaded",
          "deleted"
        ],
        "title": "VideoStatus",
        "description": "Upload lifecycle status for a video."
      },
      "VideoStreamUrlResponse": {
        "properties": {
          "video_id": {
            "type": "string",
            "format": "uuid",
            "title": "Video Id"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "stream_url"
        ],
        "title": "VideoStreamUrlResponse",
        "description": "Response with a presigned stream URL."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}